+++ conda_package_streaming/lazy_wheel.py
@@ -5,9 +5,10 @@ from __future__ import annotations
 import logging
 import zipfile
 from bisect import bisect_left, bisect_right
+from collections.abc import Iterator
 from contextlib import contextmanager
 from tempfile import NamedTemporaryFile
-from typing import Any, Iterator
+from typing import Any
 from zipfile import BadZipfile, ZipFile
 
 from requests import HTTPError, Session
@@ -49,8 +50,7 @@ class LazyZipOverHTTP:
         :param chunk_size: The chunk size to use for downloading.
         :param fall_back_to_full_download: If true, we fall back to downloading
         the whole file if the server incorrectly responds with 416
-        (Range Not Satisfiable) to an HTTP range request (Artifactory does that
-        if the file is smaller than the range requested).
+        (Range Not Satisfiable) to an HTTP range request.
         """
         # if CONTENT_CHUNK_SIZE is bigger than the file:
         # In [8]: response.headers["Content-Range"]
@@ -62,13 +62,11 @@ class LazyZipOverHTTP:
 
         self._fall_back_to_full_download: bool = fall_back_to_full_download
         self._has_streaming_support: bool = True
-        """
-        If the server returns 416 (Range Not Satisfiable) and the fallback is enabled,
-        we request the whole file and set this to False.
-        Some package servers (Artifactory) incorrectly respond with 416
-        (Range Not Satisfiable) if the file is smaller than the range requested.
-        See https://jfrog.atlassian.net/browse/RTFACT-30882
-        """
+
+        # If the server returns 416 (Range Not Satisfiable) and the fallback is
+        # enabled, we request the whole file and set this to False. Some package
+        # servers incorrectly respond with 416 (Range Not Satisfiable) if the
+        # file is smaller than the range requested.
 
         # initial range request for the end of the file
         # if the server does not support range requests, this sets
@@ -235,10 +233,9 @@ class LazyZipOverHTTP:
                 # fallback disabled, print helpful error message
                 raise HTTPError(
                     "The server returned 416 (Range Not Satisfiable). "
-                    "If you're using Artifactory, you've likely encountered a "
-                    "well-known bug in Artifactory that occurs if the file is smaller"
+                    "This can occur on a buggy server if the file is smaller "
                     "than the range requested. Set the fall_back_to_full_download flag "
-                    "to work around this issue. ",
+                    "to work around this issue.",
                     response=response,
                 )
             else: