+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 [//]: # (current developments)
 
 * Skip setting permissions if `tarinfo.mode` is `None`. (#140)
+* Set minimum Python version to 3.9.
 
 ## 0.11.0 (2024-10)